+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 = 2.5.8 =
+* Fixed: Template display for themes showing excerpts on singular pages
 * Fixed: Possible next show duplicates current show in Data API
 * Fixed: Save multiple Show related post values
 * Fixed: Count bug on Override Archive shortcode